+++ Sources/ChessKit/Clock.swift
@@ -6,7 +6,7 @@
 /// Tracks the number of moves in a game for
 /// the purposes of regulating the 50 move rule.
 ///
-public struct Clock {
+public struct Clock: Equatable {
     
     /// The maximum number of half moves before
     /// a draw by the fifty move rule should be called.