+++ Sources/ChessKit/Game.swift
@@ -5,79 +5,18 @@
 
 import Foundation
 
-/// Contains a pair of moves (white and black) and the associated turn number.
-public struct MovePair: Hashable {
-    /// The turn number of the move.
-    public var turnNumber: Int
-    /// The white move.
-    public var white: Move
-    /// The black move.
-    ///
-    /// This is optional since black goes second and may not have
-    /// made a move yet, or the game ended after white's move.
-    public var black: Move?
-    
-    /// Initialize a `MovePair` with the given moves and turn number.
-    ///
-    /// - parameter turnNumber: The number of the current move.
-    /// - parameter white: The white move in the pair.
-    /// - parameter black: The black move in the pair, if applicable.
-    ///
-    /// Each turn has a white move but not necessarily a black move
-    /// if black hasn't moved yet.
-    ///
-    public init(turnNumber: Int = 1, white: Move, black: Move? = nil) {
-        self.turnNumber = turnNumber
-        self.white = white
-        self.black = black
-    }
-    
-    /// Update the attribute of a particular move.
-    ///
-    /// - parameter color: The color of the move in the pair to update.
-    /// - parameter keyPath: The key path of the move to update.
-    /// - parameter newValue: The new value to set the key path to.
-    ///
-    public mutating func updateMove<T>(
-        with color: Piece.Color,
-        keyPath: WritableKeyPath<Move, T>,
-        newValue: T
-    ) {
-        switch color {
-        case .white:    white[keyPath: keyPath] = newValue
-        case .black:    black?[keyPath: keyPath] = newValue
-        }
-    }
-    
-    /// Returns the move within the move pair corresponding to the
-    /// provided color.
-    ///
-    /// - parameter color: The color of the requested move.
-    /// - returns: The move corresponding to `color`, or `nil`
-    /// if a move of that color doesn't exist (only applicable
-    /// for `Piece.Color.black`).
-    public func move(for color: Piece.Color) -> Move? {
-        switch color {
-        case .white:    return white
-        case .black:    return black
-        }
-    }
-}
-
 /// Represents a chess game.
 ///
 /// This object is the entry point for interacting with a full
 /// chess game within `ChessKit`. It provides methods for
 /// making moves and publishes the played moves in an observable way.
 ///
-public class Game: ObservableObject {
+public class Game: Equatable, ObservableObject {
     
-    @Published public private(set) var moves: [Int: MovePair]
-    var positions: [Position]
-    
-    var currentPosition: Position? {
-        positions.last
-    }
+    /// The move tree representing all moves made in the game.
+    @Published public private(set) var moves: MoveTree
+    /// A dictionary of every position in the game, keyed by move index.
+    public private(set) var positions: [MoveTree.Index: Position]
     
     /// Initialize a game with a starting position.
     ///
@@ -85,8 +24,8 @@ public class Game: ObservableObject {
     /// Defaults to the starting position.
     ///
     public init(startingWith position: Position = .standard) {
-        moves = [:]
-        positions = [position]
+        moves = MoveTree()
+        positions = [.minimum: position]
     }
     
     /// Initialize a game with a PGN string.
@@ -107,17 +46,38 @@ public class Game: ObservableObject {
     /// Perform the provided move in the game.
     ///
     /// - parameter move: The move to perform.
-    /// - parameter turn: The turn of the move.
-    /// - parameter color: The color of the piece that is performing the move.
+    /// - parameter index: The current move index to make the move from.
+    /// If this parameter is `nil` or omitted, the move is made from the
+    /// last move in the main variation branch.
+    /// - returns: The move index of the resulting position. If the
+    /// move couldn't be made, the provided `index` is returned directly.
     ///
-    public func make(move: Move, turn: Int, for color: Piece.Color) {
-        guard let currentPosition = positions.last else { return }
+    /// This method does not make any move legality assumptions,
+    /// it will attempt to make the move defined by `move` by moving
+    /// pieces at the provided starting/ending squares and making any
+    /// necessary captures, promotions, etc. It is the responsibility
+    /// of the caller to ensure the move is legal, see the `Board` struct.
+    ///
+    /// If `move` is the same as the upcoming move in the
+    /// current variation of `index`, the move is not made, otherwise
+    /// another variation with the same first move as the existing one
+    /// would be created.
+    ///
+    @discardableResult
+    public func make(
+        move: Move,
+        from index: MoveTree.Index
+    ) -> MoveTree.Index {
+        if let existingMoveIndex = moves.nextIndex(containing: move, for: index) {
+            // if attempted move already exists next in the variation,
+            // skip making it and return the corresponding index
+            return existingMoveIndex
+        }
         
-        switch color {
-        case .white:
-            moves[turn] = MovePair(turnNumber: turn, white: move, black: nil)
-        case .black:
-            moves[turn]?.black = move
+        let newIndex = moves.add(move: move, toParentIndex: index)
+        
+        guard let currentPosition = positions[index] else {
+            return index
         }
         
         var newPosition = currentPosition
@@ -139,7 +99,67 @@ public class Game: ObservableObject {
             newPosition.promote(pieceAt: move.end, to: promotedPiece.kind)
         }
         
-        positions.append(newPosition)
+        newPosition.toggleSideToMove()
+        positions[newIndex] = newPosition
+        return newIndex
+    }
+    
+    /// Perform the provided move in the game.
+    ///
+    /// - parameter moveString: The SAN string of the move to perform.
+    /// - parameter index: The current move index to make the move from.
+    /// If this parameter is `nil` or omitted, the move is made from the
+    /// last move in the main variation branch.
+    /// - returns: The move index of the resulting position. If the
+    /// move couldn't be made, the provided `index` is returned directly.
+    ///
+    /// This method does not make any move legality assumptions,
+    /// it will attempt to make the move defined by `moveString` by moving
+    /// pieces at the provided starting/ending squares and making any
+    /// necessary captures, promotions, etc. It is the responsibility
+    /// of the caller to ensure the move is legal, see the `Board` struct.
+    ///
+    @discardableResult
+    public func make(
+        move moveString: String,
+        from index: MoveTree.Index
+    ) -> MoveTree.Index {
+        guard let position = positions[index],
+              let move = SANParser.parse(move: moveString, in: position)
+        else {
+            return index
+        }
+        
+        return make(move: move, from: index)
+    }
+    
+    /// Perform the provided moves in the game.
+    ///
+    /// - parameter moveStrings: An array of SAN strings of the moves to perform.
+    /// - parameter index: The current move index to make the moves from.
+    /// If this parameter is `nil` or omitted, the move is made from the
+    /// last move in the main variation branch.
+    /// - returns: The move index of the resulting position. If the
+    /// moves couldn't be made, the provided `index` is returned directly.
+    ///
+    /// This method does not make any move legality assumptions,
+    /// it will attempt to make the moves defined by `moveStrings` by moving
+    /// pieces at the provided starting/ending squares and making any
+    /// necessary captures, promotions, etc. It is the responsibility
+    /// of the caller to ensure the moves are legal, see the `Board` struct.
+    ///
+    @discardableResult
+    public func make(
+        moves moveStrings: [String],
+        from index: MoveTree.Index
+    ) -> MoveTree.Index {
+        var index = index
+        
+        for moveString in moveStrings {
+            index = make(move: moveString, from: index)
+        }
+        
+        return index
     }
     
     /// The PGN represenation of the game.
@@ -147,15 +167,9 @@ public class Game: ObservableObject {
         PGNParser.convert(game: self)
     }
     
-    /// Annotate the move of a given `color` for a given `turn`.
-    public func annotate(
-        moveAt turn: Int,
-        color: Piece.Color,
-        assessment: Move.Assessment,
-        comment: String = ""
-    ) {
-        moves[turn]?.updateMove(with: color, keyPath: \.assessment, newValue: assessment)
-        moves[turn]?.updateMove(with: color, keyPath: \.comment, newValue: comment)
+    // MARK: Equatable
+    public static func == (lhs: Game, rhs: Game) -> Bool {
+        lhs.moves == rhs.moves && lhs.positions == rhs.positions
     }
     
 }