+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# [Unreleased]
+
+#### Added
+* `MoveTree` and `MoveTree.Index` objects to track move turns and variations.
+    * `Game.moves` is now a `MoveTree` object instead of `[Int: MovePair]`
+    * `MoveTree.Index` includes piece color and variation so it can be used to directly identify any single move within a game
+    * Use the properties and functions of `MoveTree` to retrieve moves within the tree as needed
+
+* `make(move:index:)` and `make(moves:index:)` with ability to make moves on `Game` with SAN strings for convenience
+    * For example: `game.make(moves: ["e4", "e5"])`
+    
+* `PGNParser.convert(game:)` now returns the PGN string for a given game, including variations.
+    * Note: `PGNParser.parse(pgn:)` still does not work with variations, this is coming in a future update.
+    
+* `Game.positions` is now public
+    * Contains a dictionary of all positions in the game by `MoveTree.Index`, including variations
+
+#### Removed
+* `Game.annotateMove`
+    * Modify `Move.assessment` and `Move.comment` directly instead
+* `MovePair`
+    * Use `Move` in conjuction with `MoveTree.Index` to track move indicies
+* `color` parameter from `SANParser.parse()`
+    * The color is now obtained from the `sideToMove` in the provided `position`
+
 # ChessKit 0.1.2
 Released Thursday, May 11, 2023.